Move all tests into a "unit" folder as they are unit tests. Also get rid of the "src" directory as all that should be tested is in "src" anyways.
