+++ tests/bootstrap.php
@@ -2,6 +2,6 @@
 set_include_path(
     sprintf('%s%s%s', realpath(__DIR__ . '/..'), PATH_SEPARATOR, get_include_path())
 );
-require 'src/autoload.php';
-require 'tests/src/Filters/FilterIteratorTestCase.php';
+require 'unit/autoload.php';
+require 'tests/unit/Filters/FilterIteratorTestCase.php';
