+++ tests/unit/Filters/FilterIteratorTestCase.php
