+++ tests/unit/Filters/NoInheritedMethodsIteratorTest.php
