+++ tests/unit/Filters/NoMagicMethodsIteratorTest.php
