+++ tests/unit/Filters/NoOldStyleConstructorIteratorTest.php
