+++ tests/unit/Filters/NoTraitMethodsIteratorTest.php
