+++ tests/unit/Filters/RegexMethodIteratorTest.php
