+++ tests/unit/Filters/_files/TestClassWithTrait.php
